oshi-core-ffm: cast operands before widening to long (Sonar S2184)
Sonar flagged nine int arithmetic operations whose result is widened to
long. No overflow is reachable at any of these sites (small constants or
single-digit values), but make the widening explicit at the operator:

  AIX     PerfstatFunctions.java:413                       (offset +8)
  Solaris SolarisCentralProcessorFFM.java:172,176          (nelem/i * Double.BYTES)
  Solaris SolarisNetworkParamsFFM.java:28,29               (HOST_NAME_MAX + 1)
  OpenBSD OpenBsdCentralProcessorFFM.java:213,214,233,234  (CP_INTR/CP_IDLE + offset)

+++ oshi-core-ffm/src/main/java/oshi/hardware/platform/unix/openbsd/OpenBsdCentralProcessorFFM.java
@@ -210,8 +210,8 @@ public class OpenBsdCentralProcessorFFM extends AbstractCentralProcessor {
                 ticks[TickType.NICE.getIndex()] = buf.getAtIndex(JAVA_LONG, CP_NICE);
                 ticks[TickType.SYSTEM.getIndex()] = buf.getAtIndex(JAVA_LONG, CP_SYS);
                 int offset = arraySize > CPUSTATES ? 1 : 0;
-                ticks[TickType.IRQ.getIndex()] = buf.getAtIndex(JAVA_LONG, CP_INTR + offset);
-                ticks[TickType.IDLE.getIndex()] = buf.getAtIndex(JAVA_LONG, CP_IDLE + offset);
+                ticks[TickType.IRQ.getIndex()] = buf.getAtIndex(JAVA_LONG, (long) CP_INTR + offset);
+                ticks[TickType.IDLE.getIndex()] = buf.getAtIndex(JAVA_LONG, (long) CP_IDLE + offset);
             }
         }
         return ticks;
@@ -230,8 +230,8 @@ public class OpenBsdCentralProcessorFFM extends AbstractCentralProcessor {
                     ticks[cpu][TickType.NICE.getIndex()] = buf.getAtIndex(JAVA_LONG, CP_NICE);
                     ticks[cpu][TickType.SYSTEM.getIndex()] = buf.getAtIndex(JAVA_LONG, CP_SYS);
                     int offset = arraySize > CPUSTATES ? 1 : 0;
-                    ticks[cpu][TickType.IRQ.getIndex()] = buf.getAtIndex(JAVA_LONG, CP_INTR + offset);
-                    ticks[cpu][TickType.IDLE.getIndex()] = buf.getAtIndex(JAVA_LONG, CP_IDLE + offset);
+                    ticks[cpu][TickType.IRQ.getIndex()] = buf.getAtIndex(JAVA_LONG, (long) CP_INTR + offset);
+                    ticks[cpu][TickType.IDLE.getIndex()] = buf.getAtIndex(JAVA_LONG, (long) CP_IDLE + offset);
                 }
             }
         }
